renaming and rearranging
