+++ configs/defaults/gatk_sv_singlesample.toml
@@ -5,42 +5,24 @@ dataset = 'seqr'
 ref_fasta = 'gs://cpg-common-main/references/hg38/v0/dragen_reference/Homo_sapiens_assembly38_masked.fasta'
 
 [images]
+cpg_workflows = "australia-southeast1-docker.pkg.dev/cpg-common/images/cpg_workflows:pr-363"
 
-cpg_workflows = "australia-southeast1-docker.pkg.dev/cpg-common/images/cpg_workflows:pr-345"
-
-# built by the GATk-SV repo
-cnmops_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/cnmops:v0.27.0"
-cnmops_virtual_env = "australia-southeast1-docker.pkg.dev/cpg-common/images/cnmops-virtual-env:v0.27.0"
-samtools_cloud_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/samtools-cloud:v0.27.0"
-samtools_cloud_virtual_env = "australia-southeast1-docker.pkg.dev/cpg-common/images/samtools-cloud-virtual-env:v0.27.0"
-sv_base_mini_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-base-mini:v0.27.0"
-sv_base_virtual_env = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-base-virtual-env:v0.27.0"
-sv_pipeline_virtual_env = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-pipeline-virtual-env:v0.27.0"
-sv_utils_env = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-utils-env:v0.27.0"
-sv_base = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-base:v0.27.0"
-sv_base_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-base:v0.27.0"
-sv_utils = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-utils:v0.27.0"
-sv_pipeline_base_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-pipeline:v0.27.0"
-sv_pipeline_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-pipeline:v0.27.0"
-sv_pipeline_qc_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-pipeline:v0.27.0"
-
-# external (not built by the gatk-sv repo)
+# GATK-SV specific
 cloud_sdk_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/cloud-sdk"
+cnmops_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/cnmops:v0.27.0"
 condense_counts_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/gatk:4.2.6.1-57-g9e03432"
-duphold_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/duphold:mw-xz-fixes-2-b1be6a9"
 gatk_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/gatk:4.2.6.1-57-g9e03432"
-gatk_docker_concordance = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/gatk:mw-sv-concordance-937c81"
-gatk_docker_pesr_override = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/gatk:4.2.6.1-57-g9e03432"
 genomes_in_the_cloud_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/genomes-in-the-cloud:2.3.2-1510681135"
-gq_recalibrator_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/gatk:0a7e1d86f"
-igv_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/igv:mw-xz-fixes-2-b1be6a9"
-manta_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/manta:5994670"
-pangenie_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/pangenie:vj-127571f"
-scramble_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/scramble:1.0.2"
-str = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/str:5994670"
-vapor_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/vapor:header-hash-2fc8f12"
-wham_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/wham:5994670"
 linux_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/ubuntu1804"
+samtools_cloud_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/samtools-cloud:v0.27.0"
+sv_base_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-base:v0.27.0"
+sv_pipeline_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-pipeline:v0.27.0"
+sv_base_mini_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv-base-mini:v0.27.0"
+
+# variant callers
+manta_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/manta:5994670"
+scramble_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/scramble:1.0.2"
+wham_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/wham:5994670"
 
 [references.broad]
 unpadded_intervals_file = 'gs://cpg-common-main/references/hg38/v0/hg38.even.handcurated.20k.intervals'
