+++ configs/gatk_sv/genotypebatch.toml
@@ -1,0 +1,3 @@
+[workflow]
+cohort_depth_vcf = []
+cohort_pesr_vcf = []
