--- configs/sandwich.toml
@@ -1,2 +1,0 @@
-[workflow]
-batch_names = ['b1b0be5754cfebd117b2a0e50b2d4087fe1af6_100']
